refactor(frontend): split SalaryStructureSection and AIAssistantSection under 500 LOC
SalaryStructureSection 661 -> 267 LOC + salary/ subfolder (4 files).
AIAssistantSection 615 -> 293 LOC + ai/ subfolder (4 files).
